requirements.txt
requests>=2.31.0
PyGithub>=1.59.1
python-docx>=0.8.11
repomix>=0.2.0
ollama>=0.1.9
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.1
rich>=13.7.0
markdown>=3.5.1
beautifulsoup4>=4.12.2
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
